fix some errors

+++ src/service/security.service.ts
@@ -1,7 +1,8 @@
 import { SessionRepository } from "../repositories/session.repository";
-import { ISession, SessionResponseType } from "../types/session.interface";
+import { ISession } from "../types/session.interface";
 import { JwtService } from "../helpers/jwtService";
 import { WithId } from "mongodb";
+import { StatusEnum } from "../types/status.enum";
 
 const repository = new SessionRepository();
 const jwtService = new JwtService();
@@ -11,8 +12,16 @@ export class SecurityService {
     return await repository.getAll();
   }
 
-  async deleteOne(deviceId: string): Promise<boolean> {
-    return await repository.deleteOne(deviceId);
+  async deleteOne(
+    deviceId: string,
+    userId: string | null,
+  ): Promise<StatusEnum> {
+    const data = await repository.findOne(deviceId);
+    if (userId && data?.userId === userId) {
+      const deleted = await repository.deleteOne(deviceId);
+      return deleted ? StatusEnum.NOT_CONTENT : StatusEnum.NOT_FOUND;
+    }
+    return StatusEnum.FORBIDDEN;
   }
 
   async deleteAll(token: string): Promise<boolean> {
